+++ huff_codegen/src/irgen/mod.rs
@@ -4,7 +4,10 @@ pub mod constants;
 /// Statement Bytecode Generation Module
 pub mod statements;
 
+/// Argument Call Module
+pub mod arg_calls;
+
 /// Prelude wraps common utilities.
 pub mod prelude {
-    pub use super::{constants::*, statements::*};
+    pub use super::{arg_calls::*, constants::*, statements::*};
 }
